signin use weixin

+++ src/js/services/tokenInterceptor.js
@@ -24,7 +24,7 @@ angular.module("Youyi").factory("tokenInterceptor", function($location, $q, loca
                 return response;
             }
             // server response
-            if (response.data.code == config.request.TOKEN_INVALID) {
+            if (response.data.code && response.data.code == config.request.TOKEN_INVALID) {
                 console.log("TOKEN_INVALID")
                 localStorageService.remove("token");
                 $location.search("uiSidebarLeft");
